+++ src/main/java/security/Checks.kt
@@ -39,6 +39,7 @@ object Checks {
     val XmlRpcServerDottedNamesCheck = CheckType("XML200", "Using allow_dotted_names option may allow attackers to execute arbitrary code.")
     val SslWrapSocketNoVersionCheck = CheckType("SSL100", "Calling wrap_socket without specifying an ssl_version could use an insecure default.")
     val SslBadProtocolsCheck = CheckType("SSL101", "Calling wrap_socket with insecure protocols.")
+    val StrFormatInspectionCheck = CheckType("STR100", "Calling format with insecure string.")
     val ShellInjectionCheck = CheckType("SH100", "Potential shell injection with unescaped input.")
     val SpawnShellInjectionCheck = CheckType("SH101", "Potentially risky call to spawned process. Check inputs for validation.")
 
@@ -53,8 +54,11 @@ object Checks {
             return "<a href='https://pycharm-security.readthedocs.io/en/latest/checks/$Code.html'>See $Code documentation</a>"
         }
 
-        fun getDescription(): String {
-            return "$this Found in '#ref'."
+        fun getDescription(includeLine: Boolean = true): String {
+            if (includeLine)
+                return "$this Found in '#ref'."
+            else
+                return "$this."
         }
 
         fun getStaticDescription(): String {